feat: many to one 조건 api에 적용

+++ board/src/main/java/com/codebene/board/service/PostService.java
@@ -1,6 +1,8 @@
 package com.codebene.board.service;
 
 import com.codebene.board.exception.post.PostNotFoundException;
+import com.codebene.board.exception.user.UserNotAllowedException;
+import com.codebene.board.model.entity.UserEntity;
 import com.codebene.board.model.post.Post;
 import com.codebene.board.model.post.PostPatchRequestBody;
 import com.codebene.board.model.post.PostPostRequestBody;
@@ -37,31 +39,38 @@ public class PostService {
     }
 
     // 게시물 생성
-    public Post createPost(PostPostRequestBody postPostRequestBody) {
-        PostEntity postEntity = new PostEntity();
-        postEntity.setBody(postPostRequestBody.body());
-
-        PostEntity savedPostEntity = postEntityRepository.save(postEntity);
+    public Post createPost(PostPostRequestBody postPostRequestBody, UserEntity currentUser) {
+        PostEntity savedPostEntity = postEntityRepository.save(PostEntity.of(postPostRequestBody.body(), currentUser));
         return Post.from(savedPostEntity);
     }
 
     // 게시물 수정
-    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody) {
+    public Post updatePost(Long postId, PostPatchRequestBody postPatchRequestBody, UserEntity currentUser) {
         PostEntity postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId));
+
+        // 작성자와 수정자 동일인 체크
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
 
         postEntity.setBody(postPatchRequestBody.body());
         PostEntity updatedPostEntity = postEntityRepository.save(postEntity);
         return Post.from(updatedPostEntity);
     }
 
-    public void deletePost(Long postId) {
+    public void deletePost(Long postId, UserEntity currentUser) {
         PostEntity postEntity = postEntityRepository
                 .findById(postId)
                 .orElseThrow(
                         () -> new PostNotFoundException(postId));
+
+        // 작성자와 수정자 동일인 체크
+        if (!postEntity.getUser().equals(currentUser)) {
+            throw new UserNotAllowedException();
+        }
 
         postEntityRepository.delete(postEntity);
     }
